feat: fix sonar
