+++ .github/workflows/build.yml
@@ -34,5 +34,6 @@ jobs:
         run: chmod +x ./gradlew
       - name: Build and analyze
         env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
-        run: ./gradlew build sonar --info
+        run: ./gradlew test jacocoTestReport sonar --info
