+++ build.gradle
@@ -60,13 +60,21 @@ sonar {
     property "sonar.projectKey", "danzolotov_sopra-fs26-group-03-server"
     property "sonar.projectName", "sopra-fs26-group-03-server"
     property "sonar.organization", "danzolotov"
+    property "sonar.coverage.jacoco.xmlReportPaths", "${layout.buildDirectory.get()}/reports/jacoco/test/jacocoTestReport.xml"
+    property "sonar.junit.reportPaths", "${layout.buildDirectory.get()}/test-results/test"
+    property "sonar.cpd.exclusions", "**/rest/dto/**/*.java,**/entity/**/*.java"
   }
 }
 
 jacocoTestReport {
     reports {
         xml.required = true
+        html.required = true
     }
+}
+
+tasks.named("sonar") {
+    dependsOn tasks.test, tasks.jacocoTestReport
 }
 
 test {
