+++ sonar-project.properties
@@ -1,3 +1,6 @@
 sonar.projectKey=danzolotov_sopra-fs26-group-03-server
 sonar.projectName=sopra-fs26-group-03-server
 sonar.organization=danzolotov
+sonar.coverage.jacoco.xmlReportPaths=build/reports/jacoco/test/jacocoTestReport.xml
+sonar.junit.reportPaths=build/test-results/test
+sonar.cpd.exclusions=**/rest/dto/**/*.java,**/entity/**/*.java
