+++ src/main/java/ch/uzh/ifi/hase/soprafs26/rest/dto/UserGetDTO.java
@@ -1,15 +1,8 @@
 package ch.uzh.ifi.hase.soprafs26.rest.dto;
 
-import ch.uzh.ifi.hase.soprafs26.constant.UserStatus;
-
-public class UserGetDTO {
+public class UserGetDTO extends BaseUserDTO {
 
 	private String userID;
-	private String email;
-	private String username;
-	private String bio;
-	private byte[] profilePicture;
-	private UserStatus status;
 
 	public String getUserID() {
 		return userID;
@@ -17,45 +10,5 @@ public class UserGetDTO {
 
 	public void setUserID(String userID) {
 		this.userID = userID;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getUsername() {
-		return username;
-	}
-
-	public void setUsername(String username) {
-		this.username = username;
-	}
-
-	public String getBio() {
-		return bio;
-	}
-
-	public void setBio(String bio) {
-		this.bio = bio;
-	}
-
-	public byte[] getProfilePicture() {
-		return profilePicture;
-	}
-
-	public void setProfilePicture(byte[] profilePicture) {
-		this.profilePicture = profilePicture;
-	}
-
-	public UserStatus getStatus() {
-		return status;
-	}
-
-	public void setStatus(UserStatus status) {
-		this.status = status;
 	}
 }
